feat(julia): PathNodeId as (gparent, parent, id) window and a whole row per UP
The UP creates one node per identifier obtained by shifting the window of the
last row, linked only to the parents that fit, with the owners its parents had.
Alias.WINDOW=2 reproduces the previous identifier: same verdicts, same node
counts on the seeded set. PathNodeId is a struct because hashing a NamedTuple
with nested Unions was ~15x slower. Row counter widened from Int8.

+++ julia/improves/src/db/path/cols/path_col_nodes.jl
@@ -7,14 +7,14 @@ module PathCollectionNodes
         step :: Step
         table :: Dict{PathNodeId, PathDocNode}
         node_ids :: SetPathNodesId
-        count :: Int8
+        count :: Int
         is_valid :: Bool
     end
 
     function new(step :: Step) :: PathColNodesLine
         table = Dict{PathNodeId, PathDocNode}()
         node_ids = SetPathNodesId()
-        count = Int8(0)
+        count = 0
         is_valid = true
         PathColNodesLine(step, table, node_ids, count, is_valid)
     end
